Created Text component. Migrated WaveHeader to SCSS
