+++ pages/register.tsx
@@ -1,48 +1,12 @@
-import { makeStyles } from "@material-ui/core"
-import { NextPage } from "next"
-import React, { ReactElement, useEffect } from "react"
-import AuthenticationDialog from "src/components/auth/AuthenticationDialog"
-import WaveHeader from "src/components/design/WaveHeader"
-import ScottyLabsIcon from "src/components/design/ScottyLabsIcon"
 import { deleteCookie } from "cookies-next"
-
-const useStyles = makeStyles((theme) => ({
-  dialog: {
-    width: "100%",
-    height: "100%",
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-    paddingTop: "10em",
-    boxSizing: "border-box",
-    [theme.breakpoints.down(theme.breakpoints.values.tablet)]: {
-      paddingTop: "3em"
-    }
-  },
-  scottyContainer: {
-    zIndex: -1,
-    opacity: 0.3,
-    bottom: 0,
-    width: "100%",
-    height: "100%",
-    position: "absolute",
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "end"
-  },
-  scottyIcon: {
-    position: "relative",
-    width: "40%",
-    bottom: 0,
-    [theme.breakpoints.down(theme.breakpoints.values.tablet)]: {
-      width: "100%"
-    }
-  }
-}))
+import { NextPage } from "next"
+import { ReactElement, useEffect } from "react"
+import AuthenticationDialog from "src/components/auth/AuthenticationDialog"
+import ScottyLabsIcon from "src/components/design/ScottyLabsIcon"
+import WaveHeader from "src/components/design/WaveHeader"
+import styles from "../styles/Register.module.scss"
 
 const RegisterPage: NextPage = (): ReactElement => {
-  const classes = useStyles()
-
   useEffect(() => {
     window.localStorage.removeItem("accessToken")
     deleteCookie("accessToken")
@@ -50,10 +14,10 @@ const RegisterPage: NextPage = (): ReactElement => {
   return (
     <div>
       <WaveHeader />
-      <div className={classes.scottyContainer}>
-        <ScottyLabsIcon className={classes.scottyIcon} />
+      <div className={styles.scottyContainer}>
+        <ScottyLabsIcon className={styles.scottyIcon} />
       </div>
-      <div className={classes.dialog}>
+      <div className={styles.dialog}>
         <AuthenticationDialog registration={true} />
       </div>
     </div>
