+++ src/components/design/Text/index.module.scss
@@ -1,0 +1,34 @@
+@import "~/styles/variables.scss";
+@import "~/styles/mixins.scss";
+
+.hero {
+  @include font-size($hero-font-sizes);
+  font-weight: 800;
+}
+
+.h1 {
+  @include font-size($h1-font-sizes);
+  font-weight: 800;
+}
+
+.h2 {
+  @include font-size($h2-font-sizes);
+  font-weight: 500;
+}
+
+.body {
+  @include font-size($body-font-sizes);
+  line-height: 1.5;
+}
+
+.subtitle {
+  @include font-size($subtitle-font-sizes);
+}
+
+.subtitle1 {
+  @include font-size($subtitle1-font-sizes);
+}
+
+.button {
+  @include font-size($button-font-sizes);
+}
