+++ styles/Register.module.scss
@@ -1,0 +1,39 @@
+@import "./variables.scss";
+@import "./mixins.scss";
+
+.dialog {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding-top: 10em;
+
+  @include down($breakpoint-md) {
+    padding-top: 3em;
+  }
+}
+
+.scottyContainer {
+  bottom: 0;
+  z-index: -1;
+  opacity: 0.3;
+  position: absolute;
+
+  width: 100%;
+  height: 100%;
+
+  display: flex;
+  justify-content: center;
+  align-items: flex-end;
+}
+
+.scottyIcon {
+  position: relative;
+  width: 40%;
+  bottom: 0;
+
+  @include down($breakpoint-md) {
+    width: 100%;
+  }
+}
