+++ styles/globals.scss
@@ -10,7 +10,7 @@ body {
   width: 100%;
   padding: 0;
   margin: 0;
-  font-family: "Rubik", sans-serif;
+  font-family: "Poppins Light", sans-serif;
   color: $text-color-primary;
   overflow-x: hidden;
   @include font-size($body-font-sizes);
