+++ styles/variables.scss
@@ -6,6 +6,7 @@ $color-primary-darker: #db4d20;
 // Text colors
 $text-color-primary: #db4d20;
 $text-color-secondary: #ffc738;
+$text-color-contrast: #fff;
 $text-color-alt: #fdcd60;
 $text-color-unfocused: #999999;
 $link-color: $color-secondary;
@@ -43,11 +44,12 @@ $hero-font-sizes: (
   xs: 2rem
 );
 $h1-font-sizes: (
-  default: 3rem
+  default: 3rem,
+  sm: 2rem
 );
 $h2-font-sizes: (
   default: 2rem,
-  xs: 1.5rem
+  sm: 1.5rem
 );
 $body-font-sizes: (
   default: 1rem
@@ -68,6 +70,7 @@ $container-padding: 1em 1.5em 1em 1.5em;
 // Shadows
 $box-shadow: 0 1px 2px rgba(0, 0, 0, 0.07), 0 2px 4px rgba(0, 0, 0, 0.07),
   0 4px 8px rgba(0, 0, 0, 0.07), 0 8px 16px rgba(0, 0, 0, 0.07);
+$drop-shadow: drop-shadow(0px 4px 8px rgba(0, 0, 0, 0.2));
 $text-shadow: 1px 1px 10px #0005;
 
 // Transitions
